move record positions to a separate file
